% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cash_Flow_Chart.R
\name{Cash_Flow_Chart}
\alias{Cash_Flow_Chart}
\title{Creates Cash Flow Chart}
\usage{
Cash_Flow_Chart(cf_table, chart_title)
}
\arguments{
\item{cf_table}{The Cash Flow Table.}

\item{chart_title}{The Chart Title.}
}
\value{
Chart for presentation.
}
\description{
This function Creates Cash Flow Chart.
}
\details{
Requires: ggplot2, dplyr
}
\examples{

}
