% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cash_Tracker_Charts.R
\name{Cash_Tracker_Charts}
\alias{Cash_Tracker_Charts}
\title{Creates Cash Tracker Charts}
\usage{
Cash_Tracker_Charts(cf_table, chart_var, chart_title)
}
\arguments{
\item{cf_table}{The Cash Flow Table.}

\item{chart_var}{The variable name of the measurement to chart.}

\item{chart_title}{The Chart Title.}
}
\value{
Chart for presentation.
}
\description{
This function Creates Cash Tracker Charts.
}
\details{
Requires: ggplot2, dplyr
}
\examples{

}
