% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merch_Bar_Charts.R
\name{Merch_Bar_Charts}
\alias{Merch_Bar_Charts}
\title{Creates Merchant Bar Chart}
\usage{
Merch_Bar_Charts(chart_data, x_axis, chart_type, chart_title)
}
\arguments{
\item{chart_data}{The Chart Data.}

\item{x_axis}{The X-Axis.}

\item{chart_type}{The Chart Type}

\item{chart_title}{The Chart Title.}
}
\value{
Chart for presentation.
}
\description{
This function Creates Merchant Bar Chart.
}
\details{
Requires: ggplot2, dplyr, scales
}
\examples{

}
